Refactor DxfClassCollection management in CadDocument
Refactored DxfClassCollection to be a read-only property of CadDocument, now requiring a reference to its parent document. Replaced static methods with instance methods, and updated AddOrUpdate and utility methods for better encapsulation. Updated DwgReader, DxfReader, and DwgClassesReader to populate the existing Classes collection instead of replacing it. Added GetInstanceCount to CadDocument for accurate class instance tracking. This improves consistency and robustness in managing DXF class definitions within a document.
